Refactor phase 2d: extract stateless timing math from GPSManager
Continues the large-file refactor tracked in
docs/development/LARGE_FILE_REFACTOR_PLAN.md. Pure motion — no behaviour
was altered.

Profiling GPSManager (2741 of gps_manager.py's 2893 lines, 54 methods) by
self usage splits it cleanly: 8 methods (386 lines) never touch instance
state, 46 (2235 lines) do. The stateless ones were @staticmethod in all
but name — pure functions trapped inside a class — and now live in:

  app_core/gps/timing_stats.py (342)  compute_jitter_summary,
                                      compute_allan_deviation,
                                      holdover_seconds, derive_leap_state
  app_core/gps/sysprobe.py     (49)   read_cpu_temp_c, safe_read

gps_manager.py drops to 2532 lines. 20 test call sites that reached
through the class (GPSManager._compute_allan_deviation(...)) now import
the functions directly. No assertion changed.

Verified as pure motion: all 6 functions are ast.dump()-identical to
their originals once the @staticmethod decorator and docstring
indentation are normalised; every non-blank removed line was asserted
present in the new modules; both implementations were run side by side
over 5 interval datasets, including empty, single-sample and constant
edge cases, with zero output differences. Full suite: 1973 passed,
67 xfailed, 0 failures.

Also corrects the plan: Phase 2 is NOT complete. The 2c pull request
claimed it was and that everything remaining was Flask-coupled or
frontend. Both gps_manager.py and app_core/radio/drivers.py are still
open and both are pure library code. They differ in kind from 2a-2c:
each is a single god-class (GPSManager 2741/2893; _SoapySDRReceiver
1822/2187), and module-level splitting cannot shrink one class. Only
their stateless parts move as verifiable motion.
