+++ app_core/gps/gps_manager.py
@@ -61,6 +61,13 @@ from pathlib import Path
 from typing import Any, Deque, Dict, List, Optional, Set, Tuple
 
 from . import ubx
+from .sysprobe import read_cpu_temp_c, safe_read
+from .timing_stats import (
+    compute_allan_deviation,
+    compute_jitter_summary,
+    derive_leap_state,
+    holdover_seconds,
+)
 
 # ---------------------------------------------------------------------------
 # clock_settime(2) helpers — used by _apply_system_time to set CLOCK_REALTIME
@@ -627,24 +634,6 @@ class GPSManager:
             except Exception as exc:  # never let a Redis hiccup drop the fix
                 self._logger.debug("Failed to persist 3D-fix anchor: %s", exc)
 
-    @staticmethod
-    def _holdover_seconds(
-        last_3d: Optional[datetime],
-        fix_mode: Any,
-        now_utc: datetime,
-    ) -> Optional[float]:
-        """Seconds since the last 3D fix for the dashboard holdover timer.
-
-        ``None`` when we have never seen a 3D fix (nothing to measure from),
-        ``0.0`` while a 3D fix is currently held, otherwise the elapsed time
-        since the anchor.  Centralised so the live ``get_status()`` view and
-        the Redis-published blob never disagree.
-        """
-        if last_3d is None:
-            return None
-        if fix_mode == 3:
-            return 0.0
-        return round((now_utc - last_3d).total_seconds(), 2)
 
     def _load_persisted_3d_fix(self) -> None:
         """Restore the holdover anchor from Redis on startup, if present."""
@@ -707,19 +696,19 @@ class GPSManager:
                 # Copy individual entries so callers can mutate the
                 # result without racing the reader-thread updates.
                 sat_history = [dict(v) for v in self._sat_history.values()]
-            self._cached_jitter = self._compute_jitter_summary(interval_samples)
+            self._cached_jitter = compute_jitter_summary(interval_samples)
             # Stability metrics ride a slower throttle — see
             # ``_adev_min_interval_s`` in __init__ for the rationale.
             if (
                 self._adev_cache_at_mono == 0.0
                 or now_mono - self._adev_cache_at_mono >= self._adev_min_interval_s
             ):
-                self._cached_allan = self._compute_allan_deviation(interval_samples)
+                self._cached_allan = compute_allan_deviation(interval_samples)
                 self._adev_cache_at_mono = now_mono
             # Host SoC temperature — sampled on the same throttle so the
             # dashboard can correlate oscillator frequency against the
             # board's thermal state (TCXO drift is mostly thermal).
-            self._cached_cpu_temp = self._read_cpu_temp_c()
+            self._cached_cpu_temp = read_cpu_temp_c()
             # Sort once at cache time so per-call output is a direct read
             # of the cached list — no per-call sort, no per-call copy.
             sat_history.sort(key=lambda e: (e.get("constellation") or "", e.get("prn") or 0))
@@ -778,7 +767,7 @@ class GPSManager:
         # ``0`` while we currently hold a 3D fix; ``None`` when we have
         # never seen one (e.g. cold start with no antenna).  Shared with
         # _publish_current_fix() so the live and Redis-cached views agree.
-        data["holdover_s"] = self._holdover_seconds(
+        data["holdover_s"] = holdover_seconds(
             last_3d, data.get("fix_mode"), now_utc
         )
         data["last_3d_fix_at"] = last_3d.isoformat() if last_3d else None
@@ -787,7 +776,7 @@ class GPSManager:
         # UBX-NAV-TIMELS poll; for now we surface a best-effort string
         # so the UI tile has something to display.  Chrony's leap_status
         # remains the authoritative source on the dashboard.
-        data["leap_state"] = self._derive_leap_state(data)
+        data["leap_state"] = derive_leap_state(data)
 
         # Per-PRN tracking history for the dashboard's almanac /
         # ephemeris staleness panel.  Cached + pre-sorted; see the
@@ -795,325 +784,6 @@ class GPSManager:
         data["satellite_history"] = self._cached_sat_history
 
         return data
-
-    @staticmethod
-    def _compute_jitter_summary(intervals_ns: List[int]) -> Dict[str, Any]:
-        """Summarise inter-pulse intervals as a histogram + scalars.
-
-        Bucket layout is adaptive: 14 inner buckets centred on zero
-        plus one under/over-flow bucket on each side (16 total).  The
-        inner bucket width is chosen from a 1-2-5 sequence so the bulk
-        of observed samples fill 5-10 buckets — that matches what
-        operators expect of a histogram and avoids the sparse 2-bar
-        appearance you get when the static ±100 µs / 20 µs grid is
-        much wider than the receiver's actual jitter.
-
-        Returns ``{}`` when the buffer holds fewer than two samples.
-        """
-        if not intervals_ns or len(intervals_ns) < 2:
-            return {
-                "sample_count": len(intervals_ns or []),
-                "histogram": [],
-                "mean_ns": None,
-                "stddev_ns": None,
-                "peak_ns": None,
-                "median_ns": None,
-            }
-
-        nominal_ns = 1_000_000_000  # 1 second
-        deltas_ns = [v - nominal_ns for v in intervals_ns]
-
-        n = len(deltas_ns)
-        mean = sum(deltas_ns) / n
-        var = sum((d - mean) * (d - mean) for d in deltas_ns) / n
-        stddev = math.sqrt(var)
-        peak = max(abs(d) for d in deltas_ns)
-        sorted_deltas = sorted(deltas_ns)
-        median = sorted_deltas[n // 2]
-
-        # Robust tail percentiles of |Δ| (nearest-rank).  σ and peak are
-        # both dominated by a handful of scheduler-latency outliers on a
-        # heavily-loaded host; p95/p99 tell the operator how bad the tail
-        # actually is without a single rogue pulse defining the headline.
-        sorted_abs = sorted(abs(d) for d in deltas_ns)
-        p95 = sorted_abs[min(n - 1, max(0, math.ceil(0.95 * n) - 1))]
-        p99 = sorted_abs[min(n - 1, max(0, math.ceil(0.99 * n) - 1))]
-
-        # Pick a bucket width so the bulk of the data spans most of the
-        # 14 inner buckets.  Target the larger of:
-        #   - half a sigma (so ±4σ fills ±8 buckets — the visible core)
-        #   - peak/14 (so a one-off outlier still lands inside the grid
-        #     rather than getting silently lumped into overflow)
-        # Then snap up to a 1-2-5 step so the X-axis tick labels stay
-        # tidy.  Floor at 100 ns to avoid zero-width buckets on
-        # exceptionally clean receivers.
-        raw_step = max(stddev / 2.0, peak / 14.0, 1.0)
-        exp10 = math.floor(math.log10(raw_step))
-        base = 10 ** exp10
-        ratio = raw_step / base
-        if ratio <= 1:
-            mult = 1
-        elif ratio <= 2:
-            mult = 2
-        elif ratio <= 5:
-            mult = 5
-        else:
-            mult = 10
-        width_ns = max(100, int(mult * base))
-
-        N_INNER = 14
-        half = N_INNER // 2  # = 7
-        edges_ns = [(i - half) * width_ns for i in range(N_INNER + 1)]
-
-        bucket_count = N_INNER + 2  # + 1 underflow, + 1 overflow
-        counts = [0] * bucket_count
-        for d in deltas_ns:
-            if d < edges_ns[0]:
-                counts[0] += 1
-            elif d >= edges_ns[-1]:
-                counts[-1] += 1
-            else:
-                # Inner-bucket index derived directly from width — no
-                # linear edge scan needed.
-                idx = (d - edges_ns[0]) // width_ns
-                if idx < 0:
-                    counts[0] += 1
-                elif idx >= N_INNER:
-                    counts[-1] += 1
-                else:
-                    counts[int(idx) + 1] += 1
-
-        def _fmt_edge(ns: int) -> str:
-            if abs(ns) >= 1000 and ns % 1000 == 0:
-                return f"{ns // 1000} µs"
-            return f"{ns} ns"
-
-        def _label(lo: Optional[int], hi: Optional[int]) -> str:
-            if lo is None:
-                return f"< {_fmt_edge(hi)}"
-            if hi is None:
-                return f"≥ {_fmt_edge(lo)}"
-            return f"{_fmt_edge(lo)} to {_fmt_edge(hi)}"
-
-        histogram: List[Dict[str, Any]] = []
-        for i, c in enumerate(counts):
-            if i == 0:
-                lo, hi = None, edges_ns[0]
-            elif i == bucket_count - 1:
-                lo, hi = edges_ns[-1], None
-            else:
-                lo, hi = edges_ns[i - 1], edges_ns[i]
-            histogram.append({
-                "label": _label(lo, hi),
-                "lo_ns": lo,
-                "hi_ns": hi,
-                "count": c,
-            })
-
-        return {
-            "sample_count": n,
-            "histogram": histogram,
-            "bucket_width_ns": width_ns,
-            "mean_ns": round(mean, 1),
-            "stddev_ns": round(stddev, 1),
-            "peak_ns": int(peak),
-            "median_ns": int(median),
-            "p95_ns": int(p95),
-            "p99_ns": int(p99),
-        }
-
-    @staticmethod
-    def _compute_allan_deviation(intervals_ns: List[int]) -> Dict[str, Any]:
-        """Overlapping Allan deviation σ_y(τ) at τ = 1, 10, 100, 1000 s.
-
-        Reconstructs the phase sequence from inter-pulse intervals
-        (x_i = cumulative interval error vs. the 1 s nominal) and
-        applies the standard overlapping ADEV estimator:
-
-            σ²_y(τ) = sum_i (x_{i+2m} - 2·x_{i+m} + x_i)²
-                      ───────────────────────────────────────
-                              2 · τ² · (N − 2m)
-
-        where m = τ / τ₀ and τ₀ = 1 s.  Returned as a flat
-        ``{tau_s: σ_y}`` dict; entries are omitted when the buffer is
-        too short to estimate ADEV at that τ (e.g. τ=1000 s needs at
-        least 2001 PPS samples in the ring).
-
-        Alongside σ_y the result carries the white-phase-modulation
-        measurement floor so consumers can tell "the oscillator is
-        wandering" apart from "the PPS timestamping chain is noisy":
-
-        * ``sigma_x_wpm_s`` — the per-pulse phase-noise estimate σ_x.
-          The interval deltas are first differences of phase, so for
-          white timestamping noise σ_x = σ_Δ / √2.
-        * ``floor_sigma_y`` — √3·σ_x/τ per returned τ (parallel array).
-          When σ_y(τ) hugs this curve the reading is measurement-noise
-          limited and says nothing about the disciplined clock itself;
-          the dashboard's stability grade uses it to avoid flagging
-          perfectly healthy clocks on hosts with µs-level PPS jitter.
-
-        Two further telecom-standard time-domain metrics ride along as
-        parallel arrays (``None`` where the buffer is too short for
-        that τ, since their sample requirements differ from ADEV's):
-
-        * ``tdev_s`` — time deviation, TDEV(τ) = τ·Mod σ_y(τ)/√3 via
-          the overlapping modified Allan variance.  This is the metric
-          the ITU-T G.811 wander masks are written against for time
-          transfer.
-        * ``mtie_s`` — maximum time interval error: the largest
-          peak-to-peak phase excursion inside any observation window of
-          length τ, computed with monotonic deques in O(N) per τ.
-        """
-        if not intervals_ns or len(intervals_ns) < 4:
-            return {
-                "tau_s": [],
-                "sigma_y": [],
-                "floor_sigma_y": [],
-                "tdev_s": [],
-                "mtie_s": [],
-                "sigma_x_wpm_s": None,
-                "sample_count": len(intervals_ns or []),
-            }
-
-        nominal_ns = 1_000_000_000
-        # Phase samples in seconds (cumulative timing error vs. ideal).
-        phase = []
-        running = 0.0
-        for v in intervals_ns:
-            running += (v - nominal_ns) / 1e9
-            phase.append(running)
-
-        n = len(phase)
-
-        # White-PM measurement floor from the interval-delta spread.
-        deltas_s = [(v - nominal_ns) / 1e9 for v in intervals_ns]
-        mean_d = sum(deltas_s) / n
-        var_d = sum((d - mean_d) * (d - mean_d) for d in deltas_s) / n
-        sigma_x = math.sqrt(var_d) / math.sqrt(2.0)
-
-        results_tau: List[int] = []
-        results_sigma: List[float] = []
-        results_floor: List[float] = []
-        results_tdev: List[Optional[float]] = []
-        results_mtie: List[Optional[float]] = []
-
-        for tau_s in (1, 10, 100, 1000):
-            m = tau_s  # τ₀ = 1 s
-            # Need at least 2m + 1 samples for a single ADEV term.
-            if n < 2 * m + 1:
-                continue
-            tau = float(m)
-            acc = 0.0
-            count = 0
-            second_diff: List[float] = []
-            for i in range(n - 2 * m):
-                d = phase[i + 2 * m] - 2.0 * phase[i + m] + phase[i]
-                second_diff.append(d)
-                acc += d * d
-                count += 1
-            if count <= 0:
-                continue
-            sigma_sq = acc / (2.0 * tau * tau * count)
-            if sigma_sq < 0:
-                continue
-            results_tau.append(tau_s)
-            results_sigma.append(math.sqrt(sigma_sq))
-            results_floor.append(math.sqrt(3.0) * sigma_x / tau)
-
-            # TDEV — overlapping modified Allan variance.  The inner
-            # m-point average of second differences is maintained as a
-            # sliding sum so the whole estimator stays O(N) per τ:
-            #
-            #   Mod σ²_y(τ) = Σ_j (Σ_{i=j}^{j+m-1} D_i)²
-            #                 ─────────────────────────────
-            #                   2 · m² · τ² · (N − 3m + 1)
-            #
-            #   TDEV(τ)     = τ · Mod σ_y(τ) / √3
-            if n >= 3 * m + 1:
-                window_sum = sum(second_diff[0:m])
-                acc_mod = window_sum * window_sum
-                terms = 1
-                for j in range(1, n - 3 * m + 1):
-                    window_sum += second_diff[j + m - 1] - second_diff[j - 1]
-                    acc_mod += window_sum * window_sum
-                    terms += 1
-                mod_avar = acc_mod / (2.0 * m * m * tau * tau * terms)
-                results_tdev.append(math.sqrt(max(0.0, mod_avar) / 3.0) * tau)
-            else:
-                results_tdev.append(None)
-
-            # MTIE — worst peak-to-peak phase excursion across every
-            # observation window of τ seconds (m+1 consecutive phase
-            # samples).  Windowed max/min tracked with monotonic deques
-            # so a 16384-sample ring stays cheap even at τ=1000 s.
-            window = m + 1
-            if n >= window:
-                max_dq: Deque[int] = deque()
-                min_dq: Deque[int] = deque()
-                mtie = 0.0
-                for idx in range(n):
-                    val = phase[idx]
-                    while max_dq and phase[max_dq[-1]] <= val:
-                        max_dq.pop()
-                    max_dq.append(idx)
-                    while min_dq and phase[min_dq[-1]] >= val:
-                        min_dq.pop()
-                    min_dq.append(idx)
-                    lo = idx - window + 1
-                    while max_dq and max_dq[0] < lo:
-                        max_dq.popleft()
-                    while min_dq and min_dq[0] < lo:
-                        min_dq.popleft()
-                    if idx >= window - 1:
-                        span = phase[max_dq[0]] - phase[min_dq[0]]
-                        if span > mtie:
-                            mtie = span
-                results_mtie.append(mtie)
-            else:
-                results_mtie.append(None)
-
-        return {
-            "tau_s": results_tau,
-            "sigma_y": results_sigma,
-            "floor_sigma_y": results_floor,
-            "tdev_s": results_tdev,
-            "mtie_s": results_mtie,
-            "sigma_x_wpm_s": sigma_x,
-            "sample_count": n,
-        }
-
-    @staticmethod
-    def _read_cpu_temp_c() -> Optional[float]:
-        """Best-effort host SoC temperature in °C from sysfs.
-
-        Prefers a thermal zone whose type mentions the CPU/SoC (covers
-        the Pi's ``cpu-thermal`` and x86's ``x86_pkg_temp``), falling
-        back to the first zone present.  Returns ``None`` on hosts
-        without a thermal zone (containers, some VMs) or on implausible
-        readings so a broken sensor can't poison the trend archive.
-        """
-        try:
-            from pathlib import Path
-            zones = sorted(Path("/sys/class/thermal").glob("thermal_zone*"))
-            chosen = None
-            for zone in zones:
-                try:
-                    ztype = (zone / "type").read_text().strip().lower()
-                except OSError:
-                    continue
-                if "cpu" in ztype or "soc" in ztype or "pkg" in ztype:
-                    chosen = zone
-                    break
-            if chosen is None and zones:
-                chosen = zones[0]
-            if chosen is None:
-                return None
-            val = float((chosen / "temp").read_text().strip()) / 1000.0
-            if -40.0 <= val <= 150.0:
-                return round(val, 1)
-        except Exception:
-            pass
-        return None
 
     # ------------------------------------------------------------------
     # Per-PRN tracking history
@@ -1194,31 +864,6 @@ class GPSManager:
                 return
         entry["last_used_at"] = now_iso
 
-    @staticmethod
-    def _derive_leap_state(fix: Dict[str, Any]) -> str:
-        """Best-effort leap-second annunciator from current fix state.
-
-        Resolution order:
-
-        1. ``UBX-NAV-TIMELS`` (Phase 2) — when ``leap_pending`` is True
-           the receiver knows about a scheduled insert/delete and we
-           surface it directly.  When it's False but ``leap_seconds`` is
-           populated, the receiver has confirmed "no event imminent"
-           and we render that as "normal".
-        2. ``has_fix`` — without UBX data, hold "normal" while we have
-           a fix so the tile isn't permanently grey.
-        3. Otherwise "unknown".
-        """
-        leap_seconds = fix.get("leap_seconds")
-        leap_pending = fix.get("leap_pending")
-        if leap_pending:
-            change = fix.get("leap_change") or 0
-            return "insert_pending" if change > 0 else "delete_pending"
-        if leap_seconds is not None:
-            return "normal"
-        if not fix.get("has_fix"):
-            return "unknown"
-        return "normal"
 
     # ------------------------------------------------------------------
     # Internal helpers
@@ -2130,7 +1775,7 @@ class GPSManager:
             # sampler's cached path) would otherwise never see holdover_s and
             # the dashboard's holdover card/tile would render blank while the
             # receiver is in holdover.
-            data["holdover_s"] = self._holdover_seconds(
+            data["holdover_s"] = holdover_seconds(
                 last_3d, data.get("fix_mode"), datetime.now(timezone.utc)
             )
             data["last_3d_fix_at"] = last_3d.isoformat() if last_3d else None
@@ -2614,12 +2259,6 @@ class GPSManager:
             self._logger.debug("Failed to scan /sys/class/pps: %s", exc)
             return None
 
-    @staticmethod
-    def _safe_read(path: Path) -> Optional[str]:
-        try:
-            return path.read_text().strip()
-        except Exception:
-            return None
 
     def _start_pps_kernel_monitor(self, device: str) -> None:
         """Spawn a low-rate poller of ``<device>/assert`` and update the fix."""
@@ -2656,7 +2295,7 @@ class GPSManager:
         ``None`` if the file is missing or the sequence is zero.
         """
         try:
-            raw = self._safe_read(Path(device) / "assert")
+            raw = safe_read(Path(device) / "assert")
             if not raw or "#" not in raw:
                 return None
             ts_part, _, seq_part = raw.partition("#")
